oh-tab-b7qq: extract git diff helpers
